structure & license, as well as some more utilities
