+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './dependencies';
+export * from './util';
