+++ tsconfig.json
@@ -7,14 +7,12 @@
 			"es6"
 		],
 		"sourceMap": true,
-		"rootDir": ".",
-		"strict": true /* enable all strict type-checking options */
+		"rootDir": "src",
+		"strict": true, /* enable all strict type-checking options */
 		/* Additional Checks */
-		// "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
-		// "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
-		// "noUnusedParameters": true,  /* Report errors on unused parameters. */
+		"noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
 	},
-	"exclude": [
-		"node_modules"
+	"include": [
+		"src"
 	]
 }
